add an aws translate vocabulary file
This helps aws translate do a slightly better job with unusual words.
The syntax is kind of painful, and their error messages are really bad,
so be very careful to test changes to this file.

Each line has 4 fields, any of which can be empty.  It must be
tab-delimited, so each line must have exactly 3 tabs, or it won't parse
correctly.

See also
https://docs.aws.amazon.com/transcribe/latest/dg/how-vocabulary.html#create-vocabulary-table