transcribe episode 014
This was transcribed by aws, and then a lot of manual cleanup was
applied.

I tried to adopt a standard-ish looking format, where a new paragraph
starts when the speaker changes.  If one speaker goes on for multiple
paragraphs, their name is only added to the first paragraph.  Speaker
names are always _italicized_, and the speaker's full name is used the
first time they speak.

There are a lot of editorial choices.  I've removed most "um" and "ah",
some of the "like, you know" (when it makes the sentences really hard to
read), etc.

If I couldn't work out what was said I inserted
  (_unintelligible: XX:XX_)
where XX:XX is the timestamp in the audio file.

I've tried to check the spelling of names and crates.  Sorry if I made
any mistakes there.

Many symbols, package names, traits, etc. have backtick quoting, though
I probably missed one or two.  In some places it's a judgement call,
e.g. is this talking about an error, or an `Error`?

In places where the speaker cuts off one thought and starts another, I
used the U+2014 "em dash" right in the source file.