show sound-bars image alongside each episode in the index
To make it more obvious where to click to go to the episode page, add a
"sound bars" image to the left of the episode title.

I'm not sure why adding "overflow: hidden" to episode_blurb messes up
the gap between episodes, but reducing the episode margin-top and
padding-top puts things more-or-less back where they were.