add a style guide for transcripts
If others start contributing transcripts, they could benefit from a set
of guidelines, so the result has consistent formatting and style.

At this point in time, this is "how things have been done so far", not
"how things must be done in the future."  There's probably a lot of
decisions in here that are short-sighted, naive, or just wrong.