add episode 009 transcript
Some of the formatting is a little tricky in this second interview.  I
think that most of the time, they're talking about projects (i.e. Tonic,
Tower) but I have used `tonic` and `tower` when we're talking about the
crates.

It's quite ambiguous, but I'm trying to not have sentences where every
third word is in `code` backticks.