simplify template logic for episodes
When the templates get evaluated, jekyll has already processed the front
matter for all pages, so we can find the transcript page by searching
for the one that identifies this page as its episode.

Switch from {% %} to {%- -%} syntax to stop it from emitting empty lines
into the episode html output. This is just to set a good example for
future editors.